make doc: API docs should show default values from source config file
...not from the user's ~/.config/stbt/stbt.conf or $STBT_CONFIG_FILE.

315fb3a0 changed `press_until_match`'s default value for `interval_secs`
and `max_presses` to be read from the configuration file. This is done
at module load time, so the config file's value is reflected in the
output of `pydoc stbt.press_until_match`.

When the user runs `make doc`, the values should come from the stbt.conf
file distributed with stb-tester, not with whatever values happen to be
in the user's configuration file at that time.