VirtualRemote.hold: Raise NotImplementedError
...instead of claiming success but not working as the user expected.

This is a YouView-internal remote-control implementation over TCP;
the listener on the set-top box doesn't seem to behave as we'd expect.

Holding the key "3" for 5 seconds with this script:

    with stbt.hold("3"):
        time.sleep(5)

does work, but only inside a text-entry box. Everywhere else, the
set-top box only sees the keypress once (as reflected by the UI, at
least). In contrast, using the above script with the LircRemote does
work when watching live TV (on the channel patch you see "3", then "33",
then "333" etc). Similarly, holding "LEFT" in the EPG only registers
once with the VirtualRemote, but works as expected with the LircRemote.