FIXUP: Remove js-doctest.py
js-doctest was added to test a single JavaScript function -
`wilson_score_interval_95`.  This is an implementation of a simple and
well known mathematical function containing only 5 lines of code and no
branches.  It it unlikely to be modified in the future or produce
incorrect results.

js-doctest was useful to validate that `wilson_score_interval_95` worked
when it was created but the cost of the test runner is too high for the
benefit it currently provides.

This commit removes the test runner and the JavaScript test for
`wilson_score_interval_95` and is a great simplification.  It should be
squashed with the previous commit before merging to master.