extra/stb-tester.spec: Relax pylint < 1.0 requirement
We want people to be able to install the stb-tester RPM on systems like
Fedora 19 that have pylint 1.0. They just won't be able to use the new
`stbt lint` tool (due to bugs with pylint, and possibly differences in
the pylint plugin API; see 97eada33 and 65a38380). Note that `stbt lint`
checks the pylint version and exits with a clear error message.