runner/templates/testrun.html: list images before log files
In failure instances, the screenshot is what you want to see first to
diagnose a failure. When collecting additional log files (using the
runner.extra-logging hook) the number of logs can push the screenshot and
other image artifacts off the bottom of the page (meaning the user must
scroll down the page for each test run they investigate.

This change simply makes the image artifacts be listed before any logs.