stbt.MatchResult: Refactor: Implement as class, not namedtuple
The user-facing Python API remains the same. I'm making this change so
that in the next commit I can add an optional argument to the
`MatchResult` constructor.

I also simplified a caller that instantiates `MatchResult` to use
positional instead of keyword arguments, to avoid unnecessary
repetition. Both the class definition and the caller are in the same
file, and anyway we shouldn't ever reorder the `MatchResult` arguments
because that would break backwards compatibility with user code in a
very confusing way.