stbt.py: Adds `black_screen` function and 4x tests
This adds a new function `black_screen` to detect the presence of a
black screen in a video frame. It converts the video frame to grayscale,
applies a binary threshold [1] with an optional mask, and then checks
that the maximum intensity value is zero. It returns `True` if it
detects a black screen and `False` otherwise.

The logic was copied from code in the existing test
`test_using_frames_to_measure_black_screen` added in commit c96503e.
However, this version adds an optional image `mask` as well as making
the threshold value configurable.

This adds new tests for `stbt.black_screen()`:
1. `test_that_black_screen_is_true_for_black_pattern`
2. `test_that_black_screen_is_false_for_smpte_pattern`
3. `test_that_black_screen_is_true_for_smpte_pattern_when_masked`
4. `test_black_screen_threshold_bounds_for_almost_black_frame`

This also adds two new images:
* `tests/videotestsrc-mask-non-black.png`: An image mask for videotestsrc
  that masks all the non-black pixels (used by test 3).
* `tests/almost-black.png`: A 720p screenshot of a typical STB black
  screen captured with a Blackmagic card (used by test 4). It is
  detected as a black screen when `threshold>=3` but not when
  `threshold<3`.

[1] OpenCV Tutorials: Basic Thresholding Operations
    http://docs.opencv.org/doc/tutorials/imgproc/threshold/threshold.html