stbt.py: Refactor mask loading into separate function
This refactors the code that loads an image mask from a file. It was
moved from `detect_motion` to a new function `_load_mask(mask)`.

This is needed in the subsequent commit, which would benefit from the
checks that the existing code performs. It would also ensure that mask
loading errors are consistent throughout stb-tester.