extra/runner/index.html: Display the plot with no passed test
jqplot has a limitation that it considers the data set passed to its
constructor invalid if plotSeries[0] is an empty array, which is the
case when no completed tests passed.

Work around this limitation by passing a fake element to plotSeries[0]
if it's otherwise empty. The timestamp (the 1st element of the fake
array) has to be valid, otherwise the placing of the elements on the
x-axis gets messed up. If 'null' is passed as the y-value (the duration
of the test), the point is not visible in the graph.