extra/runner summary report: Display the plot when no passed test
jqplot has a limitation that it considers the data set passed to its
constructor invalid if `plotSeries[0]` is an empty array, which is
the case when no completed tests passed.

Work around this limitation by passing `null` to `plotSeries[0]`
before the constructor is invoked, and removing the fake element
afterwards. It's only the constructor that has the limitation,
`plot1.replot()` is working fine if `plotSeries[0]` is empty.