stbt.black_screen: Make default `threshold` configurable in `stbt.conf`
In actual usage, the appropriate binary `threshold` value is not
expected to change from test to test. Rather, it will likely need to be
adjusted from one configuration to another (capture card, set top box,
video output type). So it makes sense to support setting the default
threshold value in `stbt.conf` itself.

This commit sets the default threshold value in both `stbt.conf` and
`tests/stbt.conf`, and updates `stbt.black_screen` to read the value
from the configuration when it is not explicitly passed as an argument.

In addition, it adds two new tests:
* `test_that_black_screen_reads_default_threshold_from_stbt_conf`
* `test_that_black_screen_threshold_parameter_overrides_default`