Rename stbt.UITest{Error,Failure} => Test{Error,Failure}
This "UI..." naming always bothered me. The old names remain for
backwards compatibility. Note that `stbt run` doesn't import the new
names into the script's namespace, because that's something we want to
move away from (we want users to explicitly "import stbt" in their
scripts). `stbt run` still imports the old names for backwards
compatibility.

I haven't tried ways of getting `raise stbt.UITestFailure()` to display
a deprecation warning, while still allowing existing scripts that catch
`UITestFailure` to catch the `MatchTimeout` (which is a `TestFailure`)
that is raised by `wait_for_match`. It might not be straightforward
enough to be worth it.