stbt.precondition: Context manager that replaces UITestFailures with UITestErrors
See the docstring of stbt.precondition for details. This approach has
proved very useful when running a single test script hundreds or
thousands of times to reproduce a particular intermittent defect. It
allows `stbt batch run` to automatically colour uninteresting† defects
in yellow (instead of red) so that we can focus our manual investigation
on the interesting failures.

† Uninteresting for the purposes of the current investigation into a
  particular defect.