Rename stbt.black_screen to stbt.is_screen_black
As pointed out by Will Manley, "black_screen" is a noun so it can be
confusing when used like this:

    black = stbt.black_screen(frame)
    [...]
    if black: ...

That is, you don't always use it directly in an `if` statement
(`if black_screen(frame): ...`); so the `is_` clarifies that this is a
predicate function. This is a policy we should adopt for all predicate
functions (of which we don't currently have any other than this
`is_screen_black` function; our functions like `wait_for_match` are
actions rather than predicates).

Aiman had originally called this `is_black_screen` but in code review I
asked him to call it `black_screen`. Sorry!

Note that renaming this function doesn't break backwards compatibility
with user scripts because the function hasn't been in any official
release of stb-tester.