Rename stbt.precondition to stbt.as_precondition
Because the following code:

    with precondition("Tuned to channel 80"):
        some_code()

...could be interpreted to mean: run `some_code()` only if the
precondition holds, or only after establishing the precondition.

What it really means is: Run `some_code()` but if it fails treat it as a
precondition error, not a test failure. This is somewhat clearer with
the new naming:

    with as_precondition("Tuned to channel 80"):
        some_code()

Note that this change doesn't break compatibility with user scripts,
because `precondition` hasn't been in any stb-tester release yet.