stbt.py: [pylint 1.0] disable certain instances of unused parameters
The functions in `Display` starting with `on_` and `restart_source`
are linked as callback to particular message types on bus, and which
must conform to certain patterns in their function signatures. The
parameters, such as _bus and _message, which are passed in are not used
by the functions, and this causes pylint to complain.

For `FakeDisplay.frames`, strictly speaking the `_timeout_secs` parameter
is *completely* unneeded, but as it is immitating the real `Display.frames`
it's nice to leave it there for better parallelism.