stbt lint: [pylint 1.0] work with either logilab.astng or astroid
Logilab renamed their `astng` module to `astroid`, but almost all of the
classes and functions that it includes have remained with the same name,
so we just include that import line in the 1.0 < pylint <= 1.0 try/except
block.

Note that I moved the E0611 (no-name-in-module) pylint warning to wrap
the entire block, and added F0401 (import-error) for `logilab.astng`.