stbt lint: '.png' does not count as a missing image file
The image path verifier should not raise an error for finding a string
that is *just* '.png' by declaring that the image does not exist on file.
It is completely understandable that users might want to use '.png' in
their code with it *not* referring to an image file. It would have to be
a hidden file, and I don't think anyone is crazy enough to use a hidden
file in their test scripts. '.png' is used as the `suffix` parameter
to a call to mktmp in stbt.py.

Note that the original regex was flawed in searching for r'.png' as this
matches any 4 character string ending with 'png' with '.' acting as a
wildcard. The new regex only returns True if it finds at least one
character followed by '.png', thus exluding simple '.png'