stbt.py, stbt-templatematch: Reword "Weak match" to "No match"
For someone who is not familiar with the template match algorithm,
the log message "Weak match found" when the match gets through
the first pass but doesn't get through the second pass is misleading.

Reword the log message to
"No match found. Closest match: MatchResult(...)"
if no match found.

The message is the same
"Match found: MatchResult(...)"
if the template matches the screenshot.