Smart TV Support: Add stbtcontraststretch element
The goal of stbt camera is to reconstruct the image sent to the TV panel
based on video recorded from a camera.  Even after careful calibration of
the camera and TV the brightness of a pixel measured at the camera may
differ from the ideal because:

* The camera suffers from vignetting (darkening at the edges)
* The TV backlight is uneven
* The received brightness of a pixel is angle dependent
* There are reflections on the screen
* Or other reasons

To deal with this location dependent brightness differences we perform
simple contrast stretching.  We take a picture with the TV displaying at
full intensity (white) and at minimum (black).  We then use these images
to scale the brightness of each individual pixel such that black becomes
black and white becomes white.

The stretching is dumb in that it assumes a linear colour-space, and is
not based on some underlying physical model.  It should probably take
Gamma into account, but it seems to work fairly well for the purposes
of image matching.  This can be revisited in the future if it turns out
to be insufficient.

This commit introduces the GStreamer element contraststretch to perform
this stretching on the video and corresponding changes to `stbt camera
calibrate` to capture the white and black images.  The core of the
implementation is written in the SIMD assembly language "orc" as used by
many GStreamer elements.  I measured this to be significantly faster than
the same algorithm implemented in C.