Smart TV Support: Add helper library for generating videos based on stills
Calibrating the camera for Smart TV support requires displaying
reference videos on the TV.  We will generate these videos on the fly
to make it easier to change them and it means we don't need to ship
100s of MB of video files with the stb-tester package.

This commit adds helper functions for generating video files based on
a series of stills.  These stills are provided as a Python sequence so
a generator of stills can be used.  This can be used to great effect
for creating dynamic videos by providing a generator which yields SVGs
modified on the fly.

`frames_to_video` generates MPEG4+AAC videos in either a mpeg-ts
container or MP4.  This is configurable because my Panasonic TV doesn't
like mpeg-ts and my Samsung TV doesn't like MP4.