Smart TV Support: Add tv driver
Calibrating the camera for Smart TV support requires displaying
reference videos on the TV.  This commit introduces an abstraction
"TV Driver", an instance of which will control the TV by some means
to display a given named video.  TV Driver generates videos on the fly
based on configuration passed in on construction (via
`tv_driver.create_from_args`).

The idea is that there might be multiple ways of getting a TV to display
a video, e.g. via DLNA or modulator support.  This commit introduces
a lowest-common-denominator approach "manual" where it's up to the user
to make the video be displayed.  We just provide them with an HTTP URL
from which to download the video.

lighttpd is used to host the videos becuase my Panasonic TV did not get
along with Python's SimpleHTTPServer.  Probably because it was single-
threaded or maybe SimpleHTTPServer doesn't support range requests.