Smart TV Support: Calibration: Add interactive colour adjustment
This calibration step is intended help adjusting the camera/TV settings
so the received pixel values are as close as possible to what we are
expecting.  It is best to do this calibration in the camera itself rather
than in software as all software adjustments are lossy, whereas adjusting
the optics/apature times, etc. in the camera before compression should be
less so.

With this commit the calibration process will show a graph of received vs.
ideal colour values and allow the user to adjust the v4l controls to make
them match as closely as possible.