stbt.wait_for_match per-frame debug output includes template filename
For long-running `wait_for_match` calls (for example with
`timeout_secs=120`) I find it really annoying to have to scroll up in
the log to find what image it's searching for. This will make the logs
much more verbose (particularly because the image path is an absolute
path†) but that's a problem that can be solved in the log-viewing tools
(turn off line-wrapping in `less`; we'll have to come up with a solution
when viewing the logs in the browser.)

† Perhaps we can truncate the path at the boundary of the git
repository, like `stbt batch run` does when reporting the test script's
filename.

`MatchResult.image` is the same thing that you passed to
`wait_for_match`: Either a string or a `numpy.ndarray`. But if it's a
string, `MatchResult.image` is slightly different than what you passed
to `wait_for_match`: `MatchResult.image` is an *absolute* path, because
you might want to use it in a different stack frame than where you
called `wait_for_match` (relative template image paths are resolved
relative to the caller of `wait_for_match` -- see `stbt._find_path` and
4e5cd23c).