Add Makefile target obs-publish
Automates pushing stbt updates to the OpenSUSE build service.  The
instructions were going to get a bunch more complicated if we wanted
to add Debian support so this commit automates that process.

This - combined with the change to fill in the rpm spec file version
automatically - also opens the door to the possibility of hooking it into
travis such that we have an stb-tester-master repo so brave souls can run
against git master without having to `make install` themselves.