test-stbt-batch.sh: Run `stbt batch run` from outside git repo
Since 00ac5b67, the selftests for `stbt batch run` create a git
repository in the temporary directory where the tests are run, to
allow the tests to pass when run from a dist tarball. `stbt batch run`
expects the scripts it's given to be in a git repository, so that it
can record the revision of the scripts it ran.

This commit changes the selftests so that `stbt batch run` is executed
from outside a git repo, but the *scripts* passed to `stbt batch run`
are inside a git repo. I'm only doing this for hypothetical reasons:
What if there was a bug in `stbt batch run` that looked at the git
revision of the current working directory instead of the script's
directory.