stbt run: Pass optional arguments after script name to script
Previously if you were to run:

    stbt run test.py --food=haloumi

stbt run would fail not recognising the optional argument `--food`.  This
commit fixes this issue.  Now all arguments that come after the test name,
optional or not, are passed to the test script.  While I believe this is
correct behaviour this change is backward incompatible.  Before if you
were to run

    stbt run test.py --source-pipeline=videotestsrc

stbt would use the source pipeline specified.  Now the source-pipeline is
passed through as an argument to the test script.  I don't think this
will be a problem in practice but it's still worth mentioning in the
release notes.

Fixes #108