stbt.py: NoVideo isn't caught by `as_precondition`
`as_precondition` is a contextmanager provided by stbt.py which allows
users to write setup steps and checks in a block in the test script, and
if any `UITestFailure` or child exceptions are raised within this block
they are caught and converted to a `PreconditionError`. This allows
`stbt batch report` to colour the row for that test run differently.

However, the catch for `UITestFailure` in `as_precondition` was also
catching `NoVideo` exceptions, which is a child of `UITestFailure`. Not
that "having video" isn't a perfectly valid precondition, but seeing
"NoVideo" as the main exception message is more useful and more important
than "PreconditionError: Didn't meet precondition <user message> (original
exception was: NoVideo)".

Therefore, I'm making NoVideo inherit from `RuntimeError` (a built-in
Python exception), which will exclude it from `as_precondition`'s catch.
This shouldn't break anyone's scripts because there shouldn't be a reason
for someone to be catching a blanket `UITestFailure` when they want to
deal with `NoVideo` - they would just catch `NoVideo`.