tests/test-stbt-batch.sh: increase timeouts for `stbt batch run` tests
I got false negatives for these tests because they were timing out
before they could conclude. This was probably due to my system having
a heavy load at the time (running tests with extra logging processes)
and the stb-tester selftests in parallel), but increase the timeout a
but won't hurt.