test_that_frames_doesnt_deadlock: Remove harmless but confusing error message
Although this self-test passes, running it with `run-tests.sh -v` showed
the following output:

    test_that_frames_doesnt_deadlock... OK
    Showing '/tmp/stb-tester.roF/log':
    ./test-stbt-py.sh: line 129: frames: command not found
    ./test-stbt-py.sh: line 129: frames: command not found

Since the offending shell command's (lack of) output ended up inside a
Python comment, it didn't affect the correctness of the test; but it was
certainly wrong and confusing.

Thanks to Lewis Haley for spotting this.