edit-testrun.js: Update table after "Failure reason" or "Notes" is edited
If "Failure reason" or "Notes" is changed in a test run's page using
`instaweb`, index.html has to be reloaded to see the updated values,
which is quite inconvenient.

This commit modifies edit-testrun.js to update the table with the
actual values straight away.

Note: the `truncate` implementation matches the behaviour of
`truncate` jinja2 filter in that `length` excludes the ellipsis,
i.e. `truncate(30)` actually returns a 33-character-long string.