edit-testrun.js: Display automatic text if manual text is deleted
If the modified "Failure reason" or "Notes" value is deleted, `instaweb`
restores the original automatic value, but the word "Empty" is
displayed in the test run's page. Get the actual value from the server's
response.

`emptytext` property of `editable` have to be redefined because its
default value is "Empty" that we wouldn't like to see.